requirements.txt
pytesseract>=0.3.10
Pillow>=10.0.0
ddgs>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0